@@ -20,13 +20,13 @@
     "ejs": "~0.8.5",
     "express": "~4.0.0",
     "express-session": "^1.15.6",
-    "install": "^0.10.2",
+    "install": "^0.10.4",
     "introjs": "^0.2.2",
     "marked": "^0.3.12",
     "morgan": "~1.0.0",
     "mysql2": "^1.4.2",
     "nodemailer": "^4.4.2",
-    "nodemon": "^1.14.8",
+    "nodemon": "^1.14.11",
     "npm": "^5.6.0",
     "prop-types": "^15.6.0",
     "react": "^16.2.0",
@@ -48,14 +48,14 @@
     "redux-persist": "^5.5.0",
     "redux-saga": "^0.16.0",
     "request": "^2.83.0",
-    "sequelize": "^4.31.2",
+    "sequelize": "^4.32.1",
     "sequelize-cli": "^3.2.0",
     "static-favicon": "~1.0.0",
     "typeface-roboto": "0.0.45"
   },
   "devDependencies": {
     "concurrently": "^3.5.1",
-    "eslint": "^4.15.0",
+    "eslint": "^4.16.0",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-plugin-compat": "^2.1.0",
     "eslint-plugin-promise": "^3.6.0"
