@@ -2,6 +2,22 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 export default class EditorCustomizeComponent extends React.Component {
+  static propTypes = {
+      changeTheme: PropTypes.func,
+      changeFontSize: PropTypes.func,
+      changeEnableBasicAutoCompletion: PropTypes.func,
+      changeEnableLiveAutoCompletion: PropTypes.func,
+      changeHighlightActiveLine: PropTypes.func
+  };
+
+  static defaultProps = {
+    changeTheme: () => {},
+    changeFontSize: () => {},
+    changeEnableBasicAutoCompletion: () => {},
+    changeEnableLiveAutoCompletion: () => {},
+    changeHighlightActiveLine: () => {}
+  };
+
   constructor(props) {
     super(props);
     this.state = {
@@ -34,12 +50,15 @@ export default class EditorCustomizeComponent extends React.Component {
           backgroundColor: '#3C444C',
           borderRight: '10px solid #343A40'
         }}>
-          <div className="nav-link form-group">
+          <div className='nav-link form-group'>
             <p style={{
               color: '#D2D3D5',
               fontFamily: 'Rubik, sans serif'
             }}>Theme</p>
-            <select className="form-control" onChange={(event)=>this.props.changeTheme(event.target.value)}>
+            <select
+              className='form-control'
+              onChange={(event)=>this.props.changeTheme(event.target.value)}
+            >
               <option>monokai</option>
               <option>github</option>
               <option>tomorrow</option>
@@ -52,16 +71,19 @@ export default class EditorCustomizeComponent extends React.Component {
               <option>terminal</option>
             </select>
           </div>
-          <div className="nav-link form-group">
+          <div className='nav-link form-group'>
             <p style={{
               color: '#D2D3D5',
               fontFamily: 'Rubik, sans serif'
             }}>Font Size</p>
-            <select className="form-control" onChange={(event)=>this.props.changeFontSize(event.target.value)}>
+            <select
+              className='form-control'
+              onChange={(event)=>this.props.changeFontSize(event.target.value)}
+            >
               <option>8</option>
               <option>10</option>
               <option>12</option>
-              <option selected="selected">14</option>
+              <option selected='selected'>14</option>
               <option>16</option>
               <option>18</option>
               <option>20</option>
@@ -71,23 +93,38 @@ export default class EditorCustomizeComponent extends React.Component {
               <option>40</option>
             </select>
           </div>
-          <div className="nav-link form-group">
-            <fieldset className="form-group">
-              <div className="form-check">
-                <label className="form-check-label">
-                  <input className="form-check-input custom-checkbox" type="checkbox" value="" onChange={(event)=>this.props.changeEnableBasicAutoCompletion(event.target.checked)}/>
+          <div className='nav-link form-group'>
+            <fieldset className='form-group'>
+              <div className='form-check'>
+                <label className='form-check-label'>
+                  <input
+                    className='form-check-input custom-checkbox'
+                    type='checkbox'
+                    value=''
+                    onChange={(event)=>this.props.changeEnableBasicAutoCompletion(event.target.checked)}
+                  />
                   <p style={{color: '#D2D3D5', fontFamily: 'Rubik, sans serif'}}>Enable Basic Autocomplete</p>
                 </label>
               </div>
-              <div className="form-check">
-                <label className="form-check-label">
-                  <input className="form-check-input" type="checkbox" value="" onChange={(event)=>this.props.changeEnableLiveAutoCompletion(event.target.checked)}/>
+              <div className='form-check'>
+                <label className='form-check-label'>
+                  <input
+                    className='form-check-input'
+                    type='checkbox'
+                    value=''
+                    onChange={(event)=>this.props.changeEnableLiveAutoCompletion(event.target.checked)}
+                  />
                   <p style={{color: '#D2D3D5', fontFamily: 'Rubik, sans serif'}}>Enable Live Autocomplete</p>
                 </label>
               </div>
-              <div className="form-check">
-                <label className="form-check-label">
-                  <input className="form-check-input" type="checkbox" value="" onChange={(event)=>this.props.changeHighlightActiveLine(event.target.checked)}/>
+              <div className='form-check'>
+                <label className='form-check-label'>
+                  <input
+                    className='form-check-input'
+                    type='checkbox'
+                    value=''
+                    onChange={(event)=>this.props.changeHighlightActiveLine(event.target.checked)}
+                  />
                   <p style={{color: '#D2D3D5', fontFamily: 'Rubik, sans serif'}}>Highlight Active Line</p>
                 </label>
               </div>
