@@ -1,21 +1,19 @@
 const API_BASE_URL = 'http://localhost:3001/code';
 
-export const codeSubmit = ({req , query}) => {
-  return fetch(API_BASE_URL + '',{
-    method: "POST",
+export const codeFetch = () => {
+  return fetch(API_BASE_URL + '/',{
+    method: "GET",
     credentials: 'include',
     headers: {
       'Accept': 'application/json',
       'Content-Type': 'application/json',
     },
-    body: JSON.stringify({
-      source: query.source
-    })
   })
     .then((response) => {
       return response.json();
     })
     .then((data) => {
+      console.log(data);
       return data;
     })
     .catch((error) => {
@@ -24,14 +22,15 @@ export const codeSubmit = ({req , query}) => {
     });
 };
 
-export const codeFetch = ({req , query}) => {
-  return fetch(API_BASE_URL + '/',{
+
+export const getCodeStatus = ({req,query}) => {
+  return fetch(API_BASE_URL + '/code_status', {
     method: "GET",
     credentials: 'include',
     headers: {
       'Accept': 'application/json',
       'Content-Type': 'application/json',
-    },
+    }
   })
     .then((response) => {
       return response.json();
@@ -58,6 +57,7 @@ export const codeLock = ({req , query}) => {
       return response.json();
     })
     .then((data) => {
+      console.log(data);
       return data;
     })
     .catch((error) => {
@@ -83,52 +83,6 @@ export const codeCompile = ({req, query}) => {
     })
     .then((data) => {
       console.log(data);
-      return data;
-    })
-    .catch((error) => {
-      console.log(error);
-      throw error;
-    });
-};
-
-export const codeExecute = ({req, query}) => {
-  return fetch(API_BASE_URL + '/', {
-    method: "POST",
-    credentials: 'include',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({
-      source: query.source
-    })
-  })
-    .then((response) => {
-      return response.json();
-    })
-    .then((data) => {
-      console.log(data);
-      return data;
-    })
-    .catch((error) => {
-      console.log(error);
-      throw error;
-    });
-};
-
-export const getCodeStatus = ({req,query}) => {
-  return fetch(API_BASE_URL + '/code_status', {
-    method: "GET",
-    credentials: 'include',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-    }
-  })
-    .then((response) => {
-      return response.json();
-    })
-    .then((data) => {
       return data;
     })
     .catch((error) => {
