@@ -1,11 +1,5 @@
 const API_BASE_URL = 'http://localhost:3001';
 
-/**
- * Promise based function to send a POST request to /login of API
- * @param req
- * @param query
- */
-
 export const userLogin = ({req , query}) => {
   return fetch(API_BASE_URL + '/user/login',{
     method: "POST",
@@ -31,32 +25,6 @@ export const userLogin = ({req , query}) => {
     });
 };
 
-export const userLoginStatus = ({req , query}) => {
-  return fetch(API_BASE_URL + '/user/login',{
-    method: "GET",
-    credentials: 'include',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-    }
-  })
-    .then((response) => {
-      return response.json();
-    })
-    .then((data) => {
-      return data;
-    })
-    .catch((error) => {
-      throw error;
-    });
-};
-
-/**
- * Promise Based function to signup
- * @param req: null
- * @param query: {emailId, name, password}
- * @returns {}
- */
 export const userRegister = ({req, query}) => {
   return fetch(API_BASE_URL + '/user/signup',{
     method: "POST",
@@ -79,6 +47,26 @@ export const userRegister = ({req, query}) => {
     })
     .catch((error) => {
       console.log(error);
+    });
+};
+
+export const userLoginStatus = ({req , query}) => {
+  return fetch(API_BASE_URL + '/user/login',{
+    method: "GET",
+    credentials: 'include',
+    headers: {
+      'Accept': 'application/json',
+      'Content-Type': 'application/json',
+    }
+  })
+    .then((response) => {
+      return response.json();
+    })
+    .then((data) => {
+      return data;
+    })
+    .catch((error) => {
+      throw error;
     });
 };
 
